Make paragraph parser more flexible
Now the chunks are not changed at all (previously commas at the end of
direct speech became periods) and the chunks of a paragraph are named
more correctly: direct (speech), indirect (speech), inquit (things like
"he said").

Since direct and indirect speech are not merged inside a paragraph
anymore, there might be more of either, so a paragraph now is a list of
tuples (type, chunk).

With the inquit still after (or before) a chunk, it should be easier to
decide which part of the direct speech it belongs to.