Try to parse run-on direct speech and detect simple quotes
Run-on direct speech is the odd style convention to not put another
quote at the start of a paragraph that still belongs to the previous
direct speech, which was never closed by a quote. Example:

"I am talking here," she said. "And I will continue to talk.
"In this paragraph it is still me talking"

Chopping up the quoted chunks now will add a little info paragraph
containing a special run-on token to indicate this happened... the
reader then can check whether the last paragraph ended with a chunk of
direct speech and then ended with a run-on token, if the next paragraph
then also starts with direct speech, then it can be merged to the last
one. This drops the information of there being a new paragraph, but the
continued direct speech should always be the more interesting
information.

Additionally:
Simple quotes are quoted phrases inside indirect speech that don't
indicate direct speech.
Example:

She wanted Jon to muss up her hair and call her "little sister" and
finish her sentences with her.

For now it is decided to simply check whether the quoted content only
contains alpha numeric chars. That's very crude, but it's not trivial to
decide this accurately. Lowercase is an approach, but proper nouns or
titles can be quoted as well.