Add script to parse out potential entities
Proper nouns or phrases of proper nouns are likely candidates for
entities such as persons, organizations, and places.

This script loads an already read and tagged JSON file and the entities
database entities.json, it enumerates all entities it can find in the
tagged data, storing it in the entities database and storing it in
generated.entities.json, so it can be reviewed and manually adjusted
before entities.json is overwritten with it.