Treat certain words as proper nouns
For instance, for GoT this is useful for "dwarf", which is used for
Tyrion a lot. Another use case is words that exist as normal English
words and therefore are hard to spot at the beginning of a sentence,
where such a word would be capitalized anyway. In GoT "Will" is an
example.