Greatly improve accuracy of inquit logic
The list of verbs to detect inquits should grow, but apart from that it
is now only the first person of the first sentence of an inquit chunk
that can override the previous direct chunk's speaker.

Furthermore, the speaker history is now stored in the chapter and there
serves as a natural concept of "persons present". The next analysing
step can use it to determine who is talking to who by simply looking
whether two names repeatedly show up consecutively.