Improve d3 graph loads
Edge values and accumulated edge values for a node now adjust the forces
applied to them, resulting in a much more realistic graph.

One also can adjust the filtering of minor edges with a slider, as well
as the default charge and link distance to play with the network a bit.

Lastly one can choose to only include linked nodes when edges have been
filtered.