Fix Arya + Bronn, Jon Snow + Grey Worm mismatch
Arya + Bronn: one of Bronn's aliases is "sellsword" and in COK Arya is
around Praed, who previously was ignored as an entity but is also called
"sellsword" in at least one sentence. So the algorithm matched that to
Bronn. Now "Praed" is added and "sellsword" is an alias for him as well,
making him the more likely match in that situation.

Jon Snow + Grey Worm: the script assumes most names to be "first name" +
"last name", so it automatically adds both words as aliases for the
person, so "Jon" and "Snow" alone are matched correctly. In the case of
"Grey Worm", this is not a good idea. In DWD there's this:

    "I hear you." The room was dim, his pallet hard. Grey light leaked
through the shutters, promising another bleak cold day.

Recognizing the uppercase "Grey" as a potential proper noun and assuming
it's "Grey Worm". "Grey Worm" now doesn't have any aliases.